# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package db

import (
	"database/sql"
	"encoding/json"
	"time"
)

type AdConfig struct {
	ID            int32
	IsEnabled     bool
	Server        string
	Port          int32
	UseTls        bool
	SkipTlsVerify bool
	BaseDn        string
	BindDn        string
	BindPassword  string
	UserFilter    string
	GroupAdminDn  string
	GroupSafetyDn string
	GroupLeaderDn string
	UpdatedAt     time.Time
	UpdatedBy     sql.NullInt64
}

type AiConfig struct {
	ID             int32
	IsEnabled      bool
	BaseUrl        string
	ApiKey         string
	DefaultModel   string
	ModelTranslate string
	ModelVision    string
	ModelSummary   string
	UpdatedAt      time.Time
	UpdatedBy      sql.NullInt64
}

type Asset struct {
	ID            int64
	SiteID        int64
	LocationCode  string
	AssetCode     string
	Name          string
	AssetType     sql.NullString
	DefaultTeamID sql.NullInt64
	IsActive      bool
	CreatedAt     time.Time
}

type CronTaskLog struct {
	ID        int64
	TaskName  string
	LastRunAt time.Time
	Status    string
	Details   sql.NullString
}

type Issue struct {
	ID                         int64
	ClientUuid                 string
	Version                    int32
	SiteID                     int64
	CreatorID                  int64
	ResolverID                 sql.NullInt64
	AssigneeID                 sql.NullInt64
	AssignedTeamID             sql.NullInt64
	Category                   string
	CauseType                  string
	VisibilityClass            string
	LocationCode               string
	AssetID                    sql.NullInt64
	CauseTeamID                sql.NullInt64
	CauseStatus                string
	Description                sql.NullString
	RejectReason               sql.NullString
	PhotoBefore                string
	PhotoDetail                sql.NullString
	PhotoAfter                 sql.NullString
	LocationNameViSnapshot     sql.NullString
	LocationNameZhSnapshot     sql.NullString
	LocationNameEnSnapshot     sql.NullString
	LocationSnapshotSource     sql.NullString
	LocationSnapshotRecordedAt sql.NullTime
	ScoreRating                sql.NullInt16
	Status                     string
	CreatedAt                  time.Time
	ResolvedAt                 sql.NullTime
	ClosedAt                   sql.NullTime
}

type IssueTag struct {
	IssueID int64
	TagCode string
}

type Location struct {
	ID        int64
	Code      string
	NameVi    string
	NameZh    string
	NameEn    string
	QrCode    string
	SiteID    int64
	IsActive  bool
	CreatedAt time.Time
}

type LocationMembership struct {
	LocationCode       string
	UserID             int64
	ResponsibilityType string
	ValidFrom          time.Time
	ValidTo            sql.NullTime
	IsActive           bool
	CreatedAt          time.Time
	CreatedBy          sql.NullInt64
}

type NotificationConfig struct {
	ID               int32
	WxpusherEnabled  bool
	WxpusherAppToken string
	LanWebhookUrl    string
	PublicBaseUrl    string
	UpdatedAt        time.Time
	UpdatedBy        sql.NullInt64
}

type NotificationOutbox struct {
	ID          int64
	IssueID     int64
	EventType   string
	Channel     string
	Payload     json.RawMessage
	Status      string
	RetryCount  int32
	MaxRetries  int32
	LastError   sql.NullString
	NextRetryAt time.Time
	CreatedAt   time.Time
	SentAt      sql.NullTime
}

type Permission struct {
	Code        string
	Description string
	IsSystem    bool
}

type RefreshToken struct {
	ID         int64
	UserID     int64
	TokenHash  string
	DeviceInfo sql.NullString
	ExpiresAt  time.Time
	RevokedAt  sql.NullTime
	CreatedAt  time.Time
}

type RolePermission struct {
	Role           string
	PermissionCode string
}

type ScoreLog struct {
	ID          int64
	IssueID     int64
	TargetType  string
	TargetID    string
	RuleKey     string
	Points      int32
	CreatedAt   time.Time
	PenaltyDate sql.NullTime
}

type ScoringRule struct {
	ID          int64
	RuleKey     string
	Points      int32
	Description sql.NullString
}

type Site struct {
	ID        int64
	Code      string
	Name      string
	Timezone  string
	IsActive  bool
	CreatedAt time.Time
}

type SystemAuditLog struct {
	ID          int64
	UserID      sql.NullInt64
	Action      string
	TargetTable string
	TargetID    string
	OldValue    json.RawMessage
	NewValue    json.RawMessage
	IpAddress   sql.NullString
	UserAgent   sql.NullString
	CreatedAt   time.Time
}

type SystemSetting struct {
	ID        int32
	Timezone  string
	UpdatedAt time.Time
	UpdatedBy sql.NullInt64
}

type Tag struct {
	ID            int64
	Code          string
	NameVi        string
	NameZh        string
	NameEn        string
	Category      string
	UseCount      int32
	IsPreset      bool
	IsActive      bool
	Status        string
	CreatedBy     sql.NullInt64
	ReviewedBy    sql.NullInt64
	ReviewedAt    sql.NullTime
	MergedTagCode sql.NullString
}

type Team struct {
	ID        int64
	SiteID    int64
	Code      string
	Name      string
	IsActive  bool
	CreatedAt time.Time
}

type TeamLocation struct {
	PeriodID     int64
	TeamID       int64
	LocationCode string
	ValidFrom    time.Time
	ValidTo      sql.NullTime
	CreatedAt    time.Time
	CreatedBy    sql.NullInt64
}

type TeamMembership struct {
	TeamID    int64
	UserID    int64
	CreatedAt time.Time
}

type TranslationCache struct {
	ContentHash    string
	TargetLang     string
	SourceText     string
	TranslatedText string
	CreatedAt      time.Time
}

type User struct {
	ID                   int64
	Username             string
	PasswordHash         sql.NullString
	AuthSource           string
	AdDn                 sql.NullString
	PinHash              sql.NullString
	BadgeCode            sql.NullString
	FullName             string
	Email                sql.NullString
	Role                 string
	SiteID               int64
	AssignedLocationCode sql.NullString
	WxUid                sql.NullString
	Timezone             sql.NullString
	Locale               string
	IsActive             bool
	CreatedAt            time.Time
	LastLoginAt          sql.NullTime
}