# internal/masterdata/handler.go
// Package masterdata provides HTTP handlers for location and tag APIs.
package masterdata

import (
	"context"
	"database/sql"
	"encoding/json"
	"net/http"
	"strings"

	"github.com/go-chi/chi/v5"

	"6s/internal/apperror"
	"6s/internal/auth"
	"6s/internal/db"
	"6s/internal/i18n"
	"6s/internal/response"
)

// Store defines repository queries for locations and tags.
type Store interface {
	ListLocations(ctx context.Context) ([]db.Location, error)
	ListAllLocations(ctx context.Context) ([]db.Location, error)
	CreateLocation(ctx context.Context, arg db.CreateLocationParams) (db.Location, error)
	UpdateLocationActiveStatus(ctx context.Context, arg db.UpdateLocationActiveStatusParams) (db.Location, error)
	UpdateLocation(ctx context.Context, arg db.UpdateLocationParams) (db.Location, error)
	ListTags(ctx context.Context) ([]db.Tag, error)
	ListAllTags(ctx context.Context) ([]db.Tag, error)
	UpsertTag(ctx context.Context, arg db.UpsertTagParams) (db.Tag, error)
	UpdateTagActiveStatus(ctx context.Context, arg db.UpdateTagActiveStatusParams) (db.Tag, error)
	SetAllTagsActiveStatus(ctx context.Context, isActive bool) error
}

type visibleTagStore interface {
	ListVisibleTags(ctx context.Context, arg db.ListVisibleTagsParams) ([]db.Tag, error)
}

type tagReviewStore interface {
	ApproveTag(ctx context.Context, arg db.ApproveTagParams) (db.Tag, error)
	RejectTag(ctx context.Context, arg db.RejectTagParams) (db.Tag, error)
	MergeTagAtomic(ctx context.Context, sourceCode, targetCode string, reviewerID int64) (db.Tag, error)
}

// Handler serves Master Data API endpoints.
type Handler struct {
	store Store
}

// NewHandler creates a new master data Handler.
func NewHandler(store Store) *Handler {
	return &Handler{store: store}
}

// LocationResponse formats location for API.
type LocationResponse struct {
	Code     string `json:"code"`
	NameVi   string `json:"name_vi"`
	NameZh   string `json:"name_zh"`
	NameEn   string `json:"name_en"`
	QRCode   string `json:"qr_code"`
	IsActive bool   `json:"is_active"`
}

// TagResponse formats tag details for API responses.
type TagResponse struct {
	Code      string `json:"code"`
	NameVi    string `json:"name_vi"`
	NameZh    string `json:"name_zh"`
	NameEn    string `json:"name_en"`
	Category  string `json:"category"`
	UseCount  int32  `json:"use_count"`
	IsPreset  bool   `json:"is_preset"`
	IsActive  bool   `json:"is_active"`
	Status    string `json:"status"`
	CreatedBy *int64 `json:"created_by"`
}

// tagResponse projects a tag row onto the API contract.
func tagResponse(t db.Tag) TagResponse {
	var createdBy *int64
	if t.CreatedBy.Valid {
		createdBy = &t.CreatedBy.Int64
	}
	return TagResponse{
		Code: t.Code, NameVi: t.NameVi, NameZh: t.NameZh, NameEn: t.NameEn, Category: t.Category,
		UseCount: t.UseCount, IsPreset: t.IsPreset, IsActive: t.IsActive, Status: t.Status, CreatedBy: createdBy,
	}
}

// ListLocations handles GET /api/locations.
func (h *Handler) ListLocations(w http.ResponseWriter, r *http.Request) {
	locs, err := h.store.ListLocations(r.Context())
	if err != nil {
		_ = response.AppError(w, r, apperror.Internal(i18n.ErrLocationQueryFailed).WithCause(err))
		return
	}

	items := make([]LocationResponse, 0, len(locs))
	for _, l := range locs {
		items = append(items, LocationResponse{
			Code:     l.Code,
			NameVi:   l.NameVi,
			NameZh:   l.NameZh,
			NameEn:   l.NameEn,
			QRCode:   l.QrCode,
			IsActive: l.IsActive,
		})
	}

	_ = response.JSON(w, http.StatusOK, items)
}

// ListAllLocations handles GET /api/locations/all (Admin only).
func (h *Handler) ListAllLocations(w http.ResponseWriter, r *http.Request) {
	locs, err := h.store.ListAllLocations(r.Context())
	if err != nil {
		_ = response.AppError(w, r, apperror.Internal(i18n.ErrLocationQueryFailed).WithCause(err))
		return
	}

	items := make([]LocationResponse, 0, len(locs))
	for _, l := range locs {
		items = append(items, LocationResponse{
			Code:     l.Code,
			NameVi:   l.NameVi,
			NameZh:   l.NameZh,
			NameEn:   l.NameEn,
			QRCode:   l.QrCode,
			IsActive: l.IsActive,
		})
	}

	_ = response.JSON(w, http.StatusOK, items)
}

// UpdateLocationStatusRequest defines payload to toggle location active status.
type UpdateLocationStatusRequest struct {
	IsActive bool `json:"is_active"`
}

// UpdateLocationStatus handles PATCH /api/locations/{code}/status (Admin only).
func (h *Handler) UpdateLocationStatus(w http.ResponseWriter, r *http.Request) {
	code := chi.URLParam(r, "code")
	if code == "" {
		code = r.PathValue("code")
	}
	if code == "" {
		code = r.URL.Query().Get("code")
	}

	var req UpdateLocationStatusRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
		return
	}

	loc, err := h.store.UpdateLocationActiveStatus(r.Context(), db.UpdateLocationActiveStatusParams{
		Code:     code,
		IsActive: req.IsActive,
	})
	if err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrLocationUpdateFailed).WithCause(err))
		return
	}

	_ = response.JSON(w, http.StatusOK, LocationResponse{
		Code:     loc.Code,
		NameVi:   loc.NameVi,
		NameZh:   loc.NameZh,
		NameEn:   loc.NameEn,
		QRCode:   loc.QrCode,
		IsActive: loc.IsActive,
	})
}

// UpdateLocationRequest defines payload to update a location's details.
type UpdateLocationRequest struct {
	NameVi string `json:"name_vi"`
	NameZh string `json:"name_zh"`
	NameEn string `json:"name_en"`
	QRCode string `json:"qr_code"`
}

// UpdateLocation handles PUT /api/locations/{code} (Admin only).
func (h *Handler) UpdateLocation(w http.ResponseWriter, r *http.Request) {
	code := chi.URLParam(r, "code")
	if code == "" {
		code = r.PathValue("code")
	}
	if code == "" {
		code = r.URL.Query().Get("code")
	}

	var req UpdateLocationRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
		return
	}

	if req.NameVi == "" || req.QRCode == "" {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrLocationMissingFields))
		return
	}

	loc, err := h.store.UpdateLocation(r.Context(), db.UpdateLocationParams{
		Code:   code,
		NameVi: req.NameVi,
		NameZh: req.NameZh,
		NameEn: req.NameEn,
		QrCode: req.QRCode,
	})
	if err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrLocationUpdateFailed).WithCause(err))
		return
	}

	_ = response.JSON(w, http.StatusOK, LocationResponse{
		Code:     loc.Code,
		NameVi:   loc.NameVi,
		NameZh:   loc.NameZh,
		NameEn:   loc.NameEn,
		QRCode:   loc.QrCode,
		IsActive: loc.IsActive,
	})
}

// CreateLocationRequest defines payload to create a new location.
type CreateLocationRequest struct {
	Code   string `json:"code"`
	NameVi string `json:"name_vi"`
	NameZh string `json:"name_zh"`
	NameEn string `json:"name_en"`
	QRCode string `json:"qr_code"`
}

// CreateLocation handles POST /api/locations (Admin only).
func (h *Handler) CreateLocation(w http.ResponseWriter, r *http.Request) {
	var req CreateLocationRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
		return
	}

	if req.Code == "" || req.NameVi == "" || req.QRCode == "" {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrLocationMissingFields))
		return
	}
	loc, err := h.store.CreateLocation(r.Context(), db.CreateLocationParams{
		Code:   req.Code,
		NameVi: req.NameVi,
		NameZh: req.NameZh,
		NameEn: req.NameEn,
		QrCode: req.QRCode,
	})
	if err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrLocationCreateFailed).WithCause(err))
		return
	}

	_ = response.JSON(w, http.StatusCreated, LocationResponse{
		Code:     loc.Code,
		NameVi:   loc.NameVi,
		NameZh:   loc.NameZh,
		NameEn:   loc.NameEn,
		QRCode:   loc.QrCode,
		IsActive: loc.IsActive,
	})
}

// ListTags handles GET /api/tags. Approved tags are public; a creator additionally sees the
// pending tags they own, and Admin/SUPERADMIN see every pending tag. Rejected and merged tags
// stay out of the picker because issue responses carry their own status details.
func (h *Handler) ListTags(w http.ResponseWriter, r *http.Request) {
	currentUser, authenticated := auth.GetUserFromContext(r.Context())
	var tags []db.Tag
	var err error
	visibleStore, hasViewerQuery := h.store.(visibleTagStore)
	if authenticated && hasViewerQuery {
		tags, err = visibleStore.ListVisibleTags(r.Context(), db.ListVisibleTagsParams{UserID: currentUser.ID, Role: currentUser.Role})
	} else {
		tags, err = h.store.ListTags(r.Context())
	}
	if err != nil {
		_ = response.AppError(w, r, apperror.Internal(i18n.ErrTagQueryFailed).WithCause(err))
		return
	}
	items := make([]TagResponse, 0, len(tags))
	for _, t := range tags {
		items = append(items, tagResponse(t))
	}

	_ = response.JSON(w, http.StatusOK, items)
}

// ListAllTags handles GET /api/tags/all (Admin only).
func (h *Handler) ListAllTags(w http.ResponseWriter, r *http.Request) {
	tags, err := h.store.ListAllTags(r.Context())
	if err != nil {
		_ = response.AppError(w, r, apperror.Internal(i18n.ErrTagQueryFailed).WithCause(err))
		return
	}

	items := make([]TagResponse, 0, len(tags))
	for _, t := range tags {
		items = append(items, tagResponse(t))
	}

	_ = response.JSON(w, http.StatusOK, items)
}

// TagReviewRequest defines an admin moderation decision for a pending tag.
type TagReviewRequest struct {
	Action        string `json:"action"`
	MergedTagCode string `json:"merged_tag_code"`
}

// ReviewTag handles PATCH /api/tags/{code}/review (Admin only).
func (h *Handler) ReviewTag(w http.ResponseWriter, r *http.Request) {
	currentUser, ok := auth.GetUserFromContext(r.Context())
	if !ok {
		_ = response.AppError(w, r, apperror.Unauthorized(i18n.ErrUnauthorized))
		return
	}
	code := strings.TrimSpace(chi.URLParam(r, "code"))
	if code == "" {
		code = strings.TrimSpace(r.PathValue("code"))
	}
	if code == "" {
		code = strings.TrimSpace(r.URL.Query().Get("code"))
	}
	if code == "" {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrTagMissingFields))
		return
	}
	reviewStore, ok := h.store.(tagReviewStore)
	if !ok {
		_ = response.AppError(w, r, apperror.Internal(i18n.ErrTagUpdateFailed))
		return
	}
	var req TagReviewRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
		return
	}
	reviewer := sql.NullInt64{Int64: currentUser.ID, Valid: currentUser.ID > 0}
	var (
		tag db.Tag
		err error
	)
	switch strings.ToUpper(strings.TrimSpace(req.Action)) {
	case "APPROVE":
		tag, err = reviewStore.ApproveTag(r.Context(), db.ApproveTagParams{Code: code, ReviewedBy: reviewer})
	case "REJECT":
		tag, err = reviewStore.RejectTag(r.Context(), db.RejectTagParams{Code: code, ReviewedBy: reviewer})
	case "MERGE":
		target := strings.TrimSpace(req.MergedTagCode)
		if target == "" {
			_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrTagMissingFields))
			return
		}
		tag, err = reviewStore.MergeTagAtomic(r.Context(), code, target, currentUser.ID)
	default:
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest, "action must be APPROVE, REJECT, or MERGE"))
		return
	}
	if err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrTagUpdateFailed).WithCause(err))
		return
	}
	type auditStore interface {
		InsertAuditLog(ctx context.Context, arg db.InsertAuditLogParams) error
	}
	if aStore, ok := h.store.(auditStore); ok {
		_ = aStore.InsertAuditLog(r.Context(), db.InsertAuditLogParams{
			UserID:      reviewer,
			Action:      "TAG_" + strings.ToUpper(strings.TrimSpace(req.Action)),
			TargetTable: "tags",
			TargetID:    code,
		})
	}
	_ = response.JSON(w, http.StatusOK, tagResponse(tag))
}

// UpdateTagStatusRequest defines payload to update tag active status.
type UpdateTagStatusRequest struct {
	IsActive bool `json:"is_active"`
}

// UpdateTagStatus handles PATCH /api/tags/{code}/status (Admin only).
func (h *Handler) UpdateTagStatus(w http.ResponseWriter, r *http.Request) {
	code := chi.URLParam(r, "code")
	if code == "" {
		code = r.PathValue("code")
	}
	if code == "" {
		code = r.URL.Query().Get("code")
	}

	var req UpdateTagStatusRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
		return
	}

	tag, err := h.store.UpdateTagActiveStatus(r.Context(), db.UpdateTagActiveStatusParams{
		Code:     code,
		IsActive: req.IsActive,
	})
	if err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrTagUpdateFailed).WithCause(err))
		return
	}

	_ = response.JSON(w, http.StatusOK, tagResponse(tag))
}

// BatchUpdateTagsStatusRequest defines payload for batch activating/deactivating tags.
type BatchUpdateTagsStatusRequest struct {
	All      *bool    `json:"all"`
	Codes    []string `json:"codes"`
	IsActive bool     `json:"is_active"`
}

// BatchUpdateTagsStatus handles POST /api/tags/batch-status (Admin only).
func (h *Handler) BatchUpdateTagsStatus(w http.ResponseWriter, r *http.Request) {
	var req BatchUpdateTagsStatusRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
		return
	}

	if req.All != nil && *req.All {
		if err := h.store.SetAllTagsActiveStatus(r.Context(), req.IsActive); err != nil {
			_ = response.AppError(w, r, apperror.Internal(i18n.ErrTagUpdateFailed).WithCause(err))
			return
		}
		_ = response.JSON(w, http.StatusOK, map[string]any{"success": true})
		return
	}

	// Update specific codes
	for _, code := range req.Codes {
		if _, err := h.store.UpdateTagActiveStatus(r.Context(), db.UpdateTagActiveStatusParams{
			Code:     code,
			IsActive: req.IsActive,
		}); err != nil {
			_ = response.AppError(w, r, apperror.Internal(i18n.ErrTagUpdateFailed).WithCause(err))
			return
		}
	}

	_ = response.JSON(w, http.StatusOK, map[string]any{"success": true})
}

// UpsertTagRequest defines payload to create or update a tag.
type UpsertTagRequest struct {
	Code     string `json:"code"`
	NameVi   string `json:"name_vi"`
	NameZh   string `json:"name_zh"`
	NameEn   string `json:"name_en"`
	Category string `json:"category"`
	IsPreset bool   `json:"is_preset"`
}

// UpsertTag handles POST /api/tags (Admin only).
func (h *Handler) UpsertTag(w http.ResponseWriter, r *http.Request) {
	var req UpsertTagRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
		return
	}

	if req.Code == "" || req.NameVi == "" || req.Category == "" {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrTagMissingFields))
		return
	}
	tag, err := h.store.UpsertTag(r.Context(), db.UpsertTagParams{
		Code:     req.Code,
		NameVi:   req.NameVi,
		NameZh:   req.NameZh,
		NameEn:   req.NameEn,
		Category: req.Category,
		IsPreset: req.IsPreset,
	})
	if err != nil {
		_ = response.AppError(w, r, apperror.Internal(i18n.ErrTagSaveFailed).WithCause(err))
		return
	}

	_ = response.JSON(w, http.StatusOK, tagResponse(tag))
}