# openapi.yaml
openapi: 3.1.0
info:
  title: 6S Issue Management & Visual Reporting System API
  version: 1.0.0
  description: API phục vụ ghi nhận, đồng bộ offline, giải quyết và tính điểm sức khỏe 6S trong môi trường nhà xưởng.
servers:
  - url: /api
    description: Runtime-configured API base path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: ISSUE_CONFLICT
        message:
          type: string
          example: Phiên bản issue đã bị thay đổi bởi người dùng khác
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    Location:
      type: object
      required:
        - code
        - name_vi
        - name_zh
        - name_en
        - qr_code
      properties:
        code:
          type: string
          example: LINE_A1
        name_vi:
          type: string
          example: Chuyền May A1
        name_zh:
          type: string
          example: 缝纫 A1 线
        name_en:
          type: string
          example: Sewing Line A1
        qr_code:
          type: string
          example: LOC:LINE_A1
        is_active:
          type: boolean
          default: true
    Tag:
      type: object
      required:
        - code
        - name_vi
        - name_zh
        - name_en
        - category
        - use_count
      properties:
        code:
          type: string
          example: oil_leak
        name_vi:
          type: string
          example: Rò rỉ dầu
        name_zh:
          type: string
          example: 漏油
        name_en:
          type: string
          example: Oil leak
        category:
          type: string
          enum: [1S, 2S, 3S, 4S, 5S, 6S]
          example: 3S
        use_count:
          type: integer
          example: 45
        is_preset:
          type: boolean
          default: false
        is_active:
          type: boolean
          default: true
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, MERGED]
          default: APPROVED
        created_by:
          type: integer
          nullable: true
...
security:
  - BearerAuth: []
paths:
...
  /locations:
    get:
      operationId: listLocations
      summary: Lấy danh mục vị trí chuẩn hóa (Master Data)
      responses:
        '200':
          description: Danh sách vị trí
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Location'
    post:
      operationId: createLocation
      summary: Thêm mới vị trí (Requires permission `masterdata:manage`)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - name_vi
                - name_zh
                - name_en
                - qr_code
              properties:
                code:
                  type: string
                  example: LINE_A3
                name_vi:
                  type: string
                  example: Chuyền May A3
                name_zh:
                  type: string
                  example: 缝纫 A3 线
                name_en:
                  type: string
                  example: Sewing Line A3
                qr_code:
                  type: string
                  example: LOC:LINE_A3
      responses:
        '201':
          description: Vị trí đã được tạo thành công
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Location'
        '403':
          description: Requires the route-specific management permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /tags:
    get:
      operationId: listTags
      summary: Lấy danh mục tags đa ngôn ngữ theo tần suất sử dụng
      responses:
        '200':
          description: Danh sách tags
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
    post:
      operationId: upsertTag
      summary: Thêm hoặc cập nhật tag (Requires permission `masterdata:manage`)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - name_vi
                - name_zh
                - name_en
                - category
              properties:
                code:
                  type: string
                  example: fire_hazard
                name_vi:
                  type: string
                  example: Nguy cơ cháy
                name_zh:
                  type: string
                  example: 火灾隐患
                name_en:
                  type: string
                  example: Fire hazard
                category:
                  type: string
                  enum: [1S, 2S, 3S, 4S, 5S, 6S]
                  example: 6S
      responses:
        '201':
          description: Tag đã được lưu thành công
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Tag'
...
  /locations/{code}/status:
    patch:
      operationId: updateLocationStatus
      summary: Bật/tắt location (Requires permission `masterdata:manage`)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cập nhật trạng thái location
          content:
            application/json:
              schema:
                type: object
  /locations/{code}:
    put:
      operationId: updateLocation
      summary: Cập nhật location (Requires permission `masterdata:manage`)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cập nhật location
          content:
            application/json:
              schema:
                type: object
...
  /tags/{code}/status:
    patch:
      operationId: updateTagStatus
      summary: Bật/tắt tag (Requires permission `masterdata:manage`)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cập nhật trạng thái tag
          content:
            application/json:
              schema:
                type: object
  /tags/{code}/review:
    patch:
      operationId: reviewTag
      summary: Duyệt, từ chối hoặc gộp tag chờ duyệt (Requires permission `masterdata:manage`)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [APPROVE, REJECT, MERGE]
                merged_tag_code:
                  type: string
                  description: Tag đích đã APPROVED khi gộp
      responses:
        '200':
          description: Kết quả cập nhật trạng thái tag
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/Tag'
...